json_video_agent/requirements.txt
google-adk
google-genai
elevenlabs
xhtml2pdf